backstage/pet_shop/pyproject.toml
[project]
name = "pet-shop-backend"
version = "0.1.0"
description = "Django backend for the Pet Marketplace System"
requires-python = ">=3.10"
dependencies = [
	"alipay-sdk-python==3.7.78",
	"asgiref==3.8.1",
	"borax==4.1.0",
	"cffi==1.16.0",
	"cryptography==42.0.5",
	"diff-match-patch==20230430",
	"django==5.0.4",
	"django-chartjs==2.3.0",
	"django-cors-headers==4.3.1",
	"django-echarts==0.6.0",
	"django-grappelli==4.0.1",
	"django-import-export==4.0.7",
	"django-jazzmin==3.0.0",
	"django-phonenumber-field==7.3.0",
	"django-simpleui==2024.4.1",
	"djangorestframework==3.15.1",
	"djangorestframework-simplejwt==5.3.1",
	"drf-yasg==1.21.7",
	"echarts-python==0.1.3",
	"et-xmlfile==1.1.0",
	"htmlgenerator==1.2.30",
	"inflection==0.5.1",
	"jinja2==3.1.4",
	"markupsafe==2.1.5",
	"pymysql==1.1.0",
	"openpyxl==3.1.4",
	"packaging==24.0",
	"phonenumbers==8.13.36",
	"pillow==10.3.0",
	"prettytable==3.10.0",
	"pyasn1==0.6.0",
	"pycparser==2.22",
	"pycryptodome==3.20.0",
	"pycryptodomex==3.20.0",
	"pyecharts==2.0.5",
	"pyjwt==2.8.0",
	"pyopenssl==24.1.0",
	"python-alipay-sdk==3.3.0",
	"pytz==2024.1",
	"pyyaml==6.0.1",
	"requests==2.32.3",
	"rsa==4.9",
	"setuptools==75.1.0",
	"simplejson==3.19.2",
	"sqlparse==0.5.0",
	"tablib==3.5.0",
	"typing_extensions==4.11.0",
	"tzdata==2024.1",
	"uritemplate==4.1.1",
	"wcwidth==0.2.13",
]
